Some cleanup and more testing
