+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/types/Type.kt
@@ -26,7 +26,6 @@
 package de.fraunhofer.aisec.cpg.graph.types
 
 import de.fraunhofer.aisec.cpg.PopulatedByPass
-import de.fraunhofer.aisec.cpg.TypeManager
 import de.fraunhofer.aisec.cpg.frontends.Language
 import de.fraunhofer.aisec.cpg.graph.Name
 import de.fraunhofer.aisec.cpg.graph.Node
@@ -38,6 +37,22 @@ import java.util.*
 import org.apache.commons.lang3.builder.ToStringBuilder
 import org.neo4j.ogm.annotation.NodeEntity
 import org.neo4j.ogm.annotation.Relationship
+
+/**
+ * This array holds the chain of different pointer/array operations. For example if a [PointerType]
+ * is built from its element type, which in turn could be a [ReferenceType] or another pointer.
+ */
+typealias TypeOperations = List<TypeOperation>
+
+/** An operation that is applied on a [Type], e.g. a pointer, an array or a reference. */
+enum class TypeOperation {
+    /** a [PointerType] with [PointerType.PointerOrigin.ARRAY] */
+    ARRAY,
+    /** a [PointerType] with [PointerType.PointerOrigin.POINTER] */
+    POINTER,
+    /** a [ReferenceType] */
+    REFERENCE,
+}
 
 /**
  * Abstract Type, describing all possible SubTypes, i.e. all different Subtypes are compliant with
@@ -190,6 +205,39 @@ abstract class Type : Node {
     }
 
     /**
+     * Calculates and returns the [TypeOperations] of the current type. A [TypeOperations] can be
+     * used to compute a "wrapped" type, for example a [PointerType] back from its [Type.root].
+     */
+    val typeOperations: TypeOperations
+        get() {
+            if (this !is SecondOrderType) {
+                return listOf()
+            }
+
+            // We already know the depth, so we can just set this and allocate the pointer origins
+            // array
+            val operations = mutableListOf<TypeOperation>()
+
+            var type: Type = this as Type
+            while (type is SecondOrderType) {
+                var op =
+                    if (type is ReferenceType) {
+                        TypeOperation.REFERENCE
+                    } else if (type is PointerType && type.isArray) {
+                        TypeOperation.ARRAY
+                    } else {
+                        TypeOperation.POINTER
+                    }
+
+                operations += op
+
+                type = type.elementType
+            }
+
+            return operations
+        }
+
+    /**
      * An ancestor is an item in a tree of types spanning from one particular [Type] to all of its
      * [Type.superTypes] (and their [Type.superTypes], and so on). Each item holds information about
      * the current "depth" within the tree.
@@ -216,6 +264,29 @@ abstract class Type : Node {
                 .toString()
         }
     }
+}
+
+/**
+ * Wraps the given [Type] into a chain of [PointerType]s and [ReferenceType]s, given the operations
+ * in [TypeOperations].
+ */
+fun TypeOperations.apply(root: Type): Type {
+    var type = root
+
+    if (this.isNotEmpty()) {
+        for (i in this.size - 1 downTo 0) {
+            var wrap = this[i]
+            if (wrap == TypeOperation.REFERENCE) {
+                type = ReferenceType(type)
+            } else if (wrap == TypeOperation.ARRAY) {
+                type = type.reference(PointerType.PointerOrigin.ARRAY)
+            } else if (wrap == TypeOperation.POINTER) {
+                type = type.reference(PointerType.PointerOrigin.POINTER)
+            }
+        }
+    }
+
+    return type
 }
 
 /** A shortcut to return [ObjectType.recordDeclaration], if this is a [ObjectType]. */
